use freemius uninstall hook, required by freemius for deployment
